23680961: models to their repositories
