+++ repository/comment/models.go
@@ -1,0 +1,8 @@
+package comment_repo
+
+type CommentItem struct {
+	IdUser  int
+	IdFilm  int
+	Rating  int
+	Comment string
+}
