+++ repository/crew/models.go
@@ -1,0 +1,8 @@
+package crew_repo
+
+type CrewItem struct {
+	Id        int
+	Name      string
+	Birthdate string
+	Photo     string
+}
