+++ repository/film/models.go
@@ -1,0 +1,11 @@
+package film_repo
+
+type FilmItem struct {
+	Id          int
+	Title       string
+	Info        string
+	Poster      string
+	ReleaseDate string
+	Country     string
+	Mpaa        string
+}
