+++ repository/genre/models.go
@@ -1,0 +1,6 @@
+package genre_repo
+
+type GenreItem struct {
+	Id    int
+	Title string
+}
