+++ repository/profession/models.go
@@ -1,0 +1,6 @@
+package profession_repo
+
+type ProfessionItem struct {
+	Id    int
+	Title string
+}
