+++ repository/user/models.go
@@ -1,0 +1,11 @@
+package user_repo
+
+type UserItem struct {
+	Id               int
+	name             string
+	Birthdate        string
+	Photo            string
+	Login            string
+	Password         string
+	RegistrationDate string
+}
